Extract shared OfflineDownloadMetaComponent type to common module
Creates offline-meta.ts to centralise the meta component type
contract, eliminating duplication between storage overview and
download card components.
